+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Change Log
 
 ### master (untagged)
 
+* Rename to serial
+
 ### 0.0.1 (2018-01-25)
 
 * Initial release